requirements.txt
azure-cognitiveservices-speech==1.38.0
azure-identity==1.17.1
black==24.4.2
isort==5.13.2
nbqa==1.8.5
python-dotenv==1.0.1
ruff~=0.1.8
